refactor: StoreService 로직 리팩터링

+++ src/main/java/eatda/controller/store/StoreInMemberResponse.java
@@ -9,7 +9,7 @@ public record StoreInMemberResponse(
         String neighborhood,
         long cheerCount
 ) {
-    public StoreInMemberResponse(Store store, int cheerCount) {
+    public StoreInMemberResponse(Store store, long cheerCount) {
         this(
                 store.getId(),
                 store.getName(),
